backend_rag/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-docx==0.8.11
PyPDF2==3.0.1
pdf2image==1.16.3
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai>=1.3.5,<2.0.0

weaviate-client==3.24.1
python-dotenv==1.0.0 